Move Config and Protocol to common-cod/
