+++ src/autov-cod/index.ts
@@ -1,5 +1,5 @@
 import { DittoCOD } from '../ditto_cod.js'
-import { Config } from './config.js'
+import { Config } from '../common-cod/config.js'
 import { HttpServer } from './http_server.js'
 
 function usage() {
